requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg==3.2.4
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
